requirements.txt
appdirs==1.4.3
autopep8==1.4.4
bcrypt==3.1.3
certifi==2019.3.9
cffi==1.10.0
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.7
django-active-users==0.3.0
django-allauth==0.39.1
django-audit-log==0.7.0
django-auditlog==0.4.3
django-cities==0.6
django-cors-headers==3.4.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==1.1.0
django-heroku==0.3.1
django-jsonfield==1.0.1
django-model-utils==4.0.0
django-mysql==3.3.0
django-phone-field==1.8.1
django-redis==4.12.1
django-sslserver==0.19
django-taggit==1.1.0
django-taggit-autocomplete-modified==0.1.0b6
djangorestframework==3.11.0
gunicorn==20.0.4
heroku==0.1.4
mysqlclient==1.4.6
oauthlib==3.0.1
packaging==16.8
pep8==1.7.1
Pillow==6.0.0
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.17
PyJWT==1.7.1
pyparsing==2.2.0
python-dateutil==1.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
redis==3.5.3
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.2.4
stripe==2.27.0
swapper==1.1.2.post1
tqdm==4.48.2
urllib3==1.24.2
whitenoise==4.1.2
